+++ engine/gamesys/src/gamesys/components/comp_sprite.cpp
@@ -30,6 +30,7 @@
 #include <dmsdk/dlib/vmath.h>
 #include <dmsdk/dlib/intersection.h>
 #include <graphics/graphics.h>
+#include <graphics/graphics_util.h>
 #include <render/render.h>
 #include <gameobject/gameobject_ddf.h>
