+++ engine/gamesys/src/gamesys/gamesys_resource.cpp
@@ -18,6 +18,7 @@
 #include "resources/res_texture.h"
 
 #include <graphics/graphics.h>
+#include <graphics/graphics_util.h>
 
 namespace dmGameSystem
 {
