+++ engine/gamesys/src/gamesys/scripts/script_resource.cpp
@@ -24,6 +24,7 @@
 #include <gamesys/texture_set_ddf.h>
 #include <graphics/graphics_ddf.h>
 #include <graphics/graphics.h>
+#include <graphics/graphics_util.h>
 #include <render/font/font_renderer.h>
 #include <resource/resource.h>
 #include <resource/resource_util.h>
