+++ engine/graphics/src/graphics_vertices.cpp
@@ -13,6 +13,7 @@
 // specific language governing permissions and limitations under the License.
 
 #include "graphics.h"
+#include "graphics_private.h"
 
 #include <dlib/math.h>
