+++ engine/render/src/render/render_script.cpp
@@ -25,6 +25,8 @@
 #include <script/script.h>
 #include <script/lua_source_ddf.h>
 
+#include <graphics/graphics_util.h>
+
 #include "font/font_renderer.h"
 #include "render/render_ddf.h"
